Python viz/requirements.txt
Flask==0.10.1
Flask-SocketIO==0.2.2
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
gevent==1.0
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.2
itsdangerous==0.23
pyserial==2.7
wsgiref==0.1.2
